Check that IntegrityError is due to a non-unique key
The code uses an IntegrityError to make sure the new database
entry has a unique key by trying to insert with force_insert=True.
However, this error can be raised if the entry values are invalid,
for instance a NULL value for field that cannot have a NULL value.
This happened when there was a typo in the field name so the correct
name was never being set.